Added and modified Lottery.cpp file related to issue #506
The lottery game shows numbers to choose and user input is checked if it matches the auto-generated random value then the user is declared winner else the program asks for retrying. The program handles if the user has mistakenly entered other numbers than the list of alphabetic characters and all.