Create Recursion-problems-hacktoberfest
I think just as all the graph problems are put into one folder all the recursion problems must be put in another folder and likewise must be done for all topics. It will not only make the repository easy to navigate but it will easy make it easy for the contributors to search for the topic they are looking for.